backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pinecone-client==2.2.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pdfplumber==0.10.3
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.26.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.10

